KAFKA-7785: move internal DefaultPartitionGrouper (#10302)
Reviewers: Guozhang Wang <wangguoz@gmail.com>
